requirements.txt
graphql-core
pywin32==304
pysimplegui==4.60.3
requests